% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fima.R
\name{dp_count}
\alias{dp_count}
\title{Differentially Private Count with Additive Laplace Mechanism}
\usage{
dp_count(count, eps = 1, delta = 1)
}
\arguments{
\item{count}{A \code{double} value (scalar or vector) representing the sample count(s) that need(s) to be privatized.}

\item{eps}{A \code{double} value representing the privacy budget \eqn{\epsilon} to guarantee \eqn{\epsilon}-DP for the resulting proportion (default value is \code{eps = 1}).}

\item{delta}{A \code{double} representing the sensitivity of the count. As a consequence the variance of the Laplace noise is \eqn{\Delta/(\epsilon)} (default value is \code{delta = 1}).}
}
\value{
A \code{double} (scalar or vector) representing the privatized count(s) guaranteeing \eqn{\epsilon}-DP (for each element in case \code{count} is a vector).
}
\description{
Based on a sample count, the function outputs a differentially private count by adding calibrated Laplace noise. More specifically, this guarantees pure DP (with privacy budget \eqn{\epsilon})
}
\examples{
\dontrun{
# Differentially private 2x2 table of independent counts
joint_probs <- c(0.25, 0.25, 0.25, 0.25) # true probabilities for multinomial
n <- 100 # total sample size
set.seed(124) # seed for reproducibility
counts <- rmultinom(1, n, prob = joint_probs) # sample counts for each category
eps <- 1 # epsilon-DP privacy budget
priv_counts <- dp_count(counts, eps, delta = 2) # DP counts for each category
priv_tab <- matrix(priv_counts, 2, 2) # Contingency table of DP counts
}
}
\author{
Roberto Molinari and Ogonnaya M. Romanus
}
