% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fima.R
\name{dp_prop}
\alias{dp_prop}
\title{Differentially Private Proportion with Additive Laplace Mechanism}
\usage{
dp_prop(prop, eps = 1, n, delta = 1)
}
\arguments{
\item{prop}{A \code{double} value (scalar or vector) representing the sample proportion(s) that need(s) to be privatized.}

\item{eps}{A \code{double} value representing the privacy budget \eqn{\epsilon} to guarantee \eqn{\epsilon}-DP for the resulting proportion (default value is \code{eps = 1}).}

\item{delta}{A \code{double} representing the sensitivity of the count associated to the proportion. As a consequence the variance of the Laplace noise is \eqn{\Delta/(n \epsilon)} (default value is \code{delta = 1}).}
}
\value{
A \code{double} (scalar or vector) representing the privatized proportion(s) guaranteeing \eqn{\epsilon}-DP (for each element in case \code{prop} is a vector).
}
\description{
Based on a sample proportion, the function outputs a differentially private proportion by adding calibrated Laplace noise. More specifically, this guarantees pure DP (with privacy budget \eqn{\epsilon})
}
\examples{
\dontrun{
# Differentially private proportion
p <- 0.8 # true probability
n <- 30 # sample size
set.seed(14) # seed for reproducibility
x <- rbinom(n, 1, prob = p) # simulate data
eps <- 1 # epsilon-DP privacy budget
pi <- dp_prop(mean(x), eps = eps, n = n) # produce DP proportion
}
}
\author{
Roberto Molinari and Ogonnaya M. Romanus
}
