% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fima.R
\name{fima_2prop}
\alias{fima_2prop}
\title{FIMA for Independent Two-Sample Proportions}
\usage{
fima_2prop(dp_stat1, dp_stat2, n1, n2, eps = 1, delta = 1, H = 10^4, seed = NA)
}
\arguments{
\item{dp_stat1}{The \code{double} value representing a privatized proportion with Laplace noise for the first sample.}

\item{dp_stat2}{The \code{double} value representing a privatized proportion with Laplace noise for the second sample.}

\item{n1}{An \code{integer} value representing the size of the first sample on which the proportion was computed.}

\item{n2}{An \code{integer} value representing the size of the second sample on which the proportion was computed.}

\item{eps}{A \code{double} value representing the privacy budget \eqn{\epsilon} used to guarantee \eqn{\epsilon}-DP for the proportions in \code{dp_stat1} and \code{dp_stat2} (default value is \code{eps = 1}).}

\item{delta}{A \code{double} representing the count-sensitivity used to guarantee \eqn{\epsilon}-DP for the proportion in \code{dp_stat} (default value is \code{delta = 1}).}

\item{H}{An \code{integer} representing the number of bootstrap solutions to approximate the FIMA distribution (default value is \code{H = 10^4}).}

\item{seed}{An \code{integer} representing a seed value to generate the random variables used in the FIMA to generate solutions (default value is \code{seed = NA}, hence no seed).}
}
\value{
A \code{vector} containing the sequence of FIMA solutions to be used for inference on the true difference in success probabilities
}
\description{
Provides the FIMA distribution for inference on the difference of true success probabilities in the independent two-sample proportions setting (see Romanus et al., 2025).
}
\examples{
\dontrun{
# FIMA distribution for two-sample proportions
p1 <- 0.8 # true success probability for first sample
p2 <- 0.9 # true success probability for second sample
n <- 100 # sample size (common for this example)
set.seed(14) # seed for reproducibility
x1 <- rbinom(n, 1, prob = p1) # simulate data for first sample
x2 <- rbinom(n, 1, prob = p2) # simulate data for second sample
pi1 <- dp_prop(mean(x1), eps = 1, n = n) # produce DP proportion for first sample
pi2 <- dp_prop(mean(x2), eps = 1, n = n) # produce DP proportion for second sample
dist <- fima_2prop(pi1, pi2, n1 = n, n2 = n, eps = eps, H = H) # produce FIMA distribution for difference in probabilities
}
}
\references{
Romanus, O.M., Boulaguiem, Y. & Molinari, R. (2025) "Fiducial Matching: Differentially Private Inference for Categorical Data".
}
\author{
Roberto Molinari and Ogonnaya M. Romanus
}
