% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fima.R
\name{fima_chi2}
\alias{fima_chi2}
\title{FIMA for \eqn{\chi^2}-test}
\usage{
fima_chi2(dp_table, n, eps = 1, delta = 2, H = 10^4, seed = NA)
}
\arguments{
\item{dp_table}{A \code{double} matrix representing the contingency table containing the privatized counts with Laplace noise.}

\item{n}{An \code{integer} value representing the sample size (i.e. total original counts).}

\item{eps}{A \code{double} value representing the privacy budget \eqn{\epsilon} used to guarantee \eqn{\epsilon}-DP for the counts in \code{dp_table} (default value is \code{eps = 1}).}

\item{delta}{A \code{double} representing the count-sensitivity used to guarantee \eqn{\epsilon}-DP for the counts in \code{dp_table} (default value is \code{delta = 2}).}

\item{H}{An \code{integer} representing the number of bootstrap solutions to approximate the FIMA distribution (default value is \code{H = 10^4}).}

\item{seed}{An \code{integer} representing a seed value to generate the random variables used in the FIMA to generate solutions (default value is \code{seed = NA}, hence no seed).}
}
\value{
A \code{vector} containing the sequence of FIMA solutions for the \eqn{\chi^2}-statistic under the null hypothesis of independence
}
\description{
Provides the FIMA distribution of the \eqn{\chi^2} statistic under the null hypothesis of independence between categorical variables (see Romanus et al., 2025).
}
\examples{
\dontrun{
# FIMA distribution for two-sample proportions
joint_probs <- c(0.25, 0.25, 0.25, 0.25) # true probabilities for multinomial
n <- 100 # total sample size
set.seed(124) # seed for reproducibility
counts <- rmultinom(1, n, prob = joint_probs) # sample counts for each category
eps <- 1 # epsilon-DP privacy budget
priv_counts <- dp_count(counts, eps, delta = 2) # DP counts for each category
priv_tab <- matrix(priv_counts, 2, 2) # Contingency table of DP counts
dist <- fima_chi2(priv_tab, n, eps) # produce FIMA distribution for chi-2 statistic under independence
}
}
\references{
Romanus, O.M., Boulaguiem, Y. & Molinari, R. (2025) "Fiducial Matching: Differentially Private Inference for Categorical Data".
}
\author{
Roberto Molinari and Ogonnaya M. Romanus
}
