% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fima.R
\name{fima_infer}
\alias{fima_infer}
\title{FIMA Inference}
\usage{
fima_infer(obj, theta0 = 0.5, ha = "greater", alpha = 0.05)
}
\arguments{
\item{obj}{The object output from the functions \code{fima_prop}, \code{fima_2prop}, \code{fima_chi2} and \code{fima_logit} (it consists in a \code{vector} representing the FIMA distribution).}

\item{theta0}{A \code{double} value representing the parameter value to test under the null hypothesis (default value is \code{theta0 = 0.5}). This is used only when \code{obj} comes from \code{fima_prop} or \code{fima_2prop}.}

\item{ha}{A \code{character} representing the one-sided alternative for the hypothesis test. The options are \code{"greater"} (i.e. we want to test if the true parameter is greater than \code{theta0}) or \code{"smaller"} (i.e. we want to test if the true parameter is smaller than \code{theta0}).}

\item{alpha}{A \code{double} value representing the significance level for the \eqn{1 - \alpha} confidence intervals on the FIMA distribution (default value is \code{alpha = 0.05}).}
}
\value{
A \code{list} containing the following objects:
\describe{
 \item{null}{The value of the parameter tested under the null hypothesis. This is produced for objects output from \code{fima_prop} and \code{fima_2prop} functions.}
 \item{alternative}{The direction of the hypothesis test under the alternative. This is produced for objects output from \code{fima_prop} and \code{fima_2prop} functions.}
 \item{p_value}{The p-value of the hypothesis test. This is produced for objects output from \code{fima_prop}, \code{fima_2prop} and \code{fima_chi2} functions.}
 \item{conf_int}{A \code{double} vector containing the lower and upper bound of the confidence intervals for the parameter of interest. This is produced for objects output from \code{fima_prop}, \code{fima_2prop} and \code{fima_logit} functions.}
}
}
\description{
A wrapper function to extract p-values (where possible) and confidence intervals (where needed) from the FIMA distributions produced by functions \code{fima_prop}, \code{fima_2prop}, \code{fima_chi2} and \code{fima_logit} (see Romanus et al., 2025).
}
\examples{
\dontrun{
# Inference for one-sample proportion
p <- 0.8 # true proportion
n <- 30 # sample size
set.seed(14) # seed for reproducibility
x <- rbinom(n, 1, prob = p) # simulate data
eps <- 1 # epsilon-DP privacy budget
pi <- dp_prop(mean(x), eps = eps, n = n) # produce DP proportion
H <- 10^4 # number of simulations for FIMA distribution
dist <- fima_prop(pi, eps = eps, n = n, H = H) # produce FIMA distribution
fima_infer(dist) # obtain inferential quantities
}
}
\references{
Romanus, O.M., Boulaguiem, Y. & Molinari, R. (2025) "Fiducial Matching: Differentially Private Inference for Categorical Data".
}
\author{
Roberto Molinari and Ogonnaya M. Romanus
}
