% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fima.R
\name{fima_logistic}
\alias{fima_logistic}
\title{FIMA for Logistic Regression (Categorical Predictors)}
\usage{
fima_logistic(dp_pi, n, eps = 1, delta = 2, H = 10^4, seed = NA)
}
\arguments{
\item{dp_pi}{A \code{double} vector representing the K proportions privatized with Laplace noise.}

\item{n}{An \code{integer} value representing the sample size (i.e. total original counts).}

\item{eps}{A \code{double} value representing the privacy budget \eqn{\epsilon} used to guarantee \eqn{\epsilon}-DP for the proportions in \code{dp_pi} (default value is \code{eps = 1}).}

\item{delta}{A \code{double} representing the count-sensitivity used to guarantee \eqn{\epsilon}-DP for the proportions in \code{dp_pi} (default value is \code{delta = 1}).}

\item{H}{An \code{integer} representing the number of bootstrap solutions to approximate the FIMA distribution (default value is \code{H = 10^4}).}

\item{seed}{An \code{integer} representing a seed value to generate the random variables used in the FIMA to generate solutions (default value is \code{seed = NA}, hence no seed).}
}
\value{
A \code{matrix} whose columns contain the FIMA distribution for each of the K coefficients from the logistic regression
}
\description{
Provides the FIMA distribution of the coefficients of logistic regression with categorical predictors. All combination of outcomes across predictors can be considered as K classes of a multinomial each with a probability of success (see Romanus et al., 2025).
}
\examples{
\dontrun{
# FIMA distribution for logistic coefficients (one categorical predictor)
joint_probs <- c(0.3, 0.2, 0.4, 0.1) # true probabilities for each combination of X = 0,1 based on T = 0,1
n <- 100 # total sample size
set.seed(124) # seed for reproducibility
prop <- rmultinom(1, n, prob = joint_probs)/n # sample proportions for each combination
eps <- 1 # epsilon-DP privacy budget
priv_props <- dp_prop(prop, n, eps, delta = 1) # DP proportions for each combination
dist <- fima_logistic(priv_props, n, eps) # produce FIMA distribution for logistic coefficients
}
}
\references{
Romanus, O.M., Boulaguiem, Y. & Molinari, R. (2025) "Fiducial Matching: Differentially Private Inference for Categorical Data".
}
\author{
Roberto Molinari and Ogonnaya M. Romanus
}
