% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fima.R
\name{fima_prop}
\alias{fima_prop}
\title{FIMA for One-Sample Proportion}
\usage{
fima_prop(dp_stat, n, eps = 1, delta = 1, H = 10^4, seed = NA)
}
\arguments{
\item{dp_stat}{The \code{double} value representing a privatized proportion with Laplace noise.}

\item{n}{A \code{double} value representing the sample size on which the non-private proportion was computed.}

\item{eps}{A \code{double} value representing the privacy budget \eqn{\epsilon} used to guarantee \eqn{\epsilon}-DP for the proportion in \code{dp_stat} (default value is \code{eps = 1}).}

\item{delta}{A \code{double} representing the count-sensitivity used to guarantee \eqn{\epsilon}-DP for the proportion in \code{dp_stat} (default value is \code{delta = 1}).}

\item{H}{An \code{integer} representing the number of bootstrap solutions to approximate the FIMA distribution (default value is \code{H = 10^4}).}

\item{seed}{An \code{integer} representing a seed value to generate the random variables used in the FIMA to generate solutions (default value is \code{seed = NA}, hence no seed).}
}
\value{
A \code{vector} containing the sequence of FIMA solutions to be used for inference on the true success probability
}
\description{
Provides the FIMA distribution for inference on the true success probability in the one-sample proportion setting (see Romanus et al., 2025).
}
\examples{
\dontrun{
# FIMA distribution for one-sample proportion
p <- 0.8 # true proportion
n <- 30 # sample size
set.seed(14) # seed for reproducibility
x <- rbinom(n, 1, prob = p) # simulate data
eps <- 1 # epsilon-DP privacy budget
pi <- dp_prop(mean(x), eps = eps, n = n) # produce DP proportion
H <- 10^4 # number of simulations for FIMA distribution
dist <- fima_prop(pi, eps = eps, n = n, H = H) # produce FIMA distribution
}
}
\references{
Romanus, O.M., Boulaguiem, Y. & Molinari, R. (2025) "Fiducial Matching: Differentially Private Inference for Categorical Data".
}
\author{
Roberto Molinari and Ogonnaya M. Romanus
}
